Move ChatWithHistoryPluginOptions and change some documentation
